RegisterInfo split up and moved.
