+++ device/src/DataDescriptor.cpp
@@ -1,44 +1,5 @@
-#include "RegisterInfo.h"
-
-namespace ChimeraTK {
-  RegisterPath RegisterInfo::getRegisterName() const{
-    return impl->getRegisterName();
-  }
-
-  unsigned int RegisterInfo::getNumberOfElements() const{
-    return impl->getNumberOfElements();
-  }
-
-  unsigned int RegisterInfo::getNumberOfChannels() const{
-    return impl->getNumberOfChannels();
-  }
-
-  unsigned int RegisterInfo::getNumberOfDimensions() const{
-    return impl->getNumberOfDimensions();
-  }
-
-  const DataDescriptor& RegisterInfo::getDataDescriptor() const{
-    return impl->getDataDescriptor();
-  }
-
-  bool RegisterInfo::isReadable() const{
-    return impl->isReadable();
-  }
-
-  bool RegisterInfo::isWriteable() const{
-    return impl->isWriteable();
-  }
-
-  AccessModeFlags RegisterInfo::getSupportedAccessModes() const{
-    return impl->getSupportedAccessModes();
-  }
-
-  boost::shared_ptr<RegisterInfoImpl> RegisterInfo::getImpl(){
-      return impl;
-  }
-
-  /*******************************************************************************************************************/
-
+#include "DataDescriptor.h"
+namespace ChimeraTK{
   DataDescriptor::DataDescriptor(DataType type) {
     if(type.isNumeric()) {
       _fundamentalType = FundamentalType::numeric;
@@ -86,5 +47,4 @@ namespace ChimeraTK {
       _nFractionalDigits = 325;
     }
   }
-
 }
